refactor(frontend): extract match tracking hook
